AbnormalClassification/requirements.txt
absl-py==0.11.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
dataclasses==0.8
efficientnet-pytorch==0.7.0
future==0.18.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.10
importlib-metadata==2.0.0
jsonpatch==1.26
jsonpointer==2.0
Markdown==3.3.3
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
Pillow==4.0.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
torch==1.7.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.8.1
tornado==6.1
tqdm==4.53.0
typing-extensions==3.7.4.3
urllib3==1.26.2
visdom==0.1.8.9
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.4.0
ml-collections==0.1.0